Read file handler
