+++ src/rpc/mod.rs
@@ -4,6 +4,8 @@
 pub mod server;
 pub mod packets;
 mod codec;
+mod fs;
+mod nfs_program;
 
 pub mod events {
     use super::packets::{
